+++ src/Lib/Exceptions/ControllerNotValidException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Symphony\ApiFramework\Lib\Exceptions;
+
+class ControllerNotValidException extends \Exception
+{
+}